fixes 'refactor gcp, minikube, docker-for-desktop, ack to be kfctl plugins'
