+++ bootstrap/Makefile
@@ -47,7 +47,10 @@ build-kfctl: generate fmt vet
 	$(GO) build -i -gcflags 'all=-N -l' -o bin/kfctl cmd/kfctl/main.go
 
 build-foo-plugin: generate fmt vet
-	$(GO) build -i -gcflags 'all=-N -l' -o bin/foo.so -buildmode=plugin cmd/plugins/foo.go
+	$(GO) build -i -gcflags 'all=-N -l' -o bin/foo.so -buildmode=plugin cmd/plugins/foo/foo.go
+
+build-minikube-plugin: generate fmt vet
+	$(GO) build -i -gcflags 'all=-N -l' -o bin/minikube.so -buildmode=plugin cmd/plugins/minikube/minikube.go
 
 build-local: build-bootstrap build-kfctl
 
@@ -81,11 +84,11 @@ run-local-docker:
 	  --entrypoint /bin/bash $(IMG):$(TAG)
 
 # init ~/myapp --platform none
-test-known-platforms-init: install build-foo-plugin
+test-known-platforms-init: install build-foo-plugin build-minikube-plugin
 	@rm -rf $(HOME)/ksonnet && \
 	kfctl init $(HOME)/ksonnet -V --platform none --version master && \
 	rm -rf $(HOME)/minikube && \
-	kfctl init $(HOME)/minikube -V --platform minikube --version master && \
+	PLUGINS_ENVIRONMENT=$(GOPATH)/src/github.com/kubeflow/kubeflow/bootstrap/bin kfctl init $(HOME)/minikube -V --platform minikube --version master && \
 	rm -rf $(HOME)/foo && \
 	PLUGINS_ENVIRONMENT=$(GOPATH)/src/github.com/kubeflow/kubeflow/bootstrap/bin kfctl init $(HOME)/foo -V --platform foo --version master && \
 	echo SUCCESS
@@ -95,11 +98,7 @@ test-known-platforms-generate: test-known-platforms-init
 	@cd ~/ksonnet && \
 	kfctl generate all -V && \
 	cd ~/minikube && \
-	kfctl generate all -V --mount-local && \
+	PLUGINS_ENVIRONMENT=$(GOPATH)/src/github.com/kubeflow/kubeflow/bootstrap/bin kfctl generate all -V --mount-local && \
 	cd ~/foo && \
 	PLUGINS_ENVIRONMENT=$(GOPATH)/src/github.com/kubeflow/kubeflow/bootstrap/bin kfctl generate all -V && \
 	echo SUCCESS
-
-test-foo-plugin:
-	rm -rf $(HOME)/foo
-	PLUGINS_ENVIRONMENT=$(GOPATH)/src/github.com/kubeflow/kubeflow/bootstrap/bin kfctl init $(HOME)/foo --platform foo
