+++ bootstrap/pkg/client/foo/foo.go
@@ -1,32 +1,19 @@
-package main
+package foo
 
 import (
 	"fmt"
 	kftypes "github.com/kubeflow/kubeflow/bootstrap/pkg/apis/apps"
-	"github.com/kubeflow/kubeflow/bootstrap/pkg/client/ksonnet"
 )
 
 // Foo implements KfApp Interface
 // It includes ksonnet along with additional Foo types
 type Foo struct {
-	ksonnet kftypes.KfApp
+	Ksonnet kftypes.KfApp
 	//TODO add additional types required for foo platform
 }
 
-func GetKfApp(options map[string]interface{}) kftypes.KfApp {
-	_foo := &Foo{
-		ksonnet: ksonnet.GetKfApp(options),
-	}
-	return _foo
-}
-
-func (foo *Foo) writeConfigFile() error {
-	//TODO write files under AppDir
-	return nil
-}
-
 func (foo *Foo) Apply(resources kftypes.ResourceEnum, options map[string]interface{}) error {
-	ksonnetlyErr := foo.ksonnet.Apply(resources, options)
+	ksonnetlyErr := foo.Ksonnet.Apply(resources, options)
 	if ksonnetlyErr != nil {
 		return fmt.Errorf("foo apply failed for ksonnet: %v", ksonnetlyErr)
 	}
@@ -34,7 +21,7 @@ func (foo *Foo) Apply(resources kftypes.ResourceEnum, options map[string]interfa
 }
 
 func (foo *Foo) Delete(resources kftypes.ResourceEnum, options map[string]interface{}) error {
-	ksDeleteErr := foo.ksonnet.Delete(resources, options)
+	ksDeleteErr := foo.Ksonnet.Delete(resources, options)
 	if ksDeleteErr != nil {
 		return fmt.Errorf("foo delete failed for ksonnet: %v", ksDeleteErr)
 	}
@@ -42,7 +29,7 @@ func (foo *Foo) Delete(resources kftypes.ResourceEnum, options map[string]interf
 }
 
 func (foo *Foo) Generate(resources kftypes.ResourceEnum, options map[string]interface{}) error {
-	ksGenerateErr := foo.ksonnet.Generate(resources, options)
+	ksGenerateErr := foo.Ksonnet.Generate(resources, options)
 	if ksGenerateErr != nil {
 		return fmt.Errorf("foo generate failed for ksonnet: %v", ksGenerateErr)
 	}
@@ -50,7 +37,7 @@ func (foo *Foo) Generate(resources kftypes.ResourceEnum, options map[string]inte
 }
 
 func (foo *Foo) Init(options map[string]interface{}) error {
-	ksInitErr := foo.ksonnet.Init(options)
+	ksInitErr := foo.Ksonnet.Init(options)
 	if ksInitErr != nil {
 		return fmt.Errorf("foo init failed for ksonnet: %v", ksInitErr)
 	}
